reafactor the structure

+++ src/Modules/Adapters/Adapter.Aero/Model/Metadata/AeroMetadata.cs
@@ -5,10 +5,10 @@ namespace Adapter.Aero.Model;
 public sealed class AeroMetadata
 {
       public bool PortOne { get; set; }
-      public int ProtocolOne { get; set; }
-      public int BaudRateOne { get; set; }
+      public short ProtocolOne { get; set; }
+      public short BaudRateOne { get; set; }
       public bool PortTwo { get; set; }
-      public int ProtocolTwo { get; set; }
-      public int BaudRateTwo { get; set; }
+      public short ProtocolTwo { get; set; }
+      public short BaudRateTwo { get; set; }
 }
