+++ src/Modules/Contracts/Device.Contract/Command/ModuleUpdateCommand.cs
@@ -3,4 +3,4 @@ using SharedKernel.Messaging;
 
 namespace Device.Contract.Command;
 
-public sealed record ModuleUpdateCommand(string Mac,int Id,string SerialNumber,string Fw,int Port) : ICommand;
+public sealed record ModuleUpdateCommand(string Mac,int Id,string SerialNumber,string Fw,short Port) : ICommand;
